some fixes to formatting and work on integrating vector db

+++ src/engine/result.hpp
@@ -8,5 +8,5 @@ struct Result
     std::string name;
     std::filesystem::path icon;
     std::filesystem::path executable;
-    double score; // Score of a result: [0, 1] 1 is ideal
+    float score; // Score of a result: [0, 1] 1 is ideal
 };
